pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.7"
openai = "0.28"
opgg-py = "^1.4.1"
apscheduler = "^3.10.4"
aiogram = "^3.14.0"
telepot = "^12.7"
telebot = "^0.0.5"
selenium = "^4.26.1"
pillow = "^11.0.0"
webdriver-manager = "^4.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
